Replace debounce with xhr.abort to shorten time-to-render
It turns out that the debounce was taking more time (150ms) than the
average round trip to the dev recommendation server with a good network
connection (~100ms). We want to ensure stale results are never shown to
the user, but we can do that more efficiently by ensuring there's always
at most one request in flight.